Refactor: Enhance OOD, Utils, DTOs; Incomplete Status/HasStatus Update
This commit includes several significant refactoring efforts aimed at improving the object-oriented design, centralizing utilities, and clarifying data transfer objects.

Completed Refactorings:
- Core processing logic (data source, partition, batch) moved from YAFPF.java
  into dedicated processor classes (DataSourceProcessor, PartitionProcessor,
  BatchProcessor).
- MetricsManager created to centralize all metrics-related records (Status,
  HasStatus, ProcessingResult, etc.) and helper methods.
- Dummy data generation moved to TestDataGenerator in test sources.
- YAFPF.execute() simplified to orchestrate calls to new processors.
- Configuration management refactored: ConfigManager.main and MetricSimulator
  moved to test/simulation; configuration records organized into individual
  files; runtime state records moved to RuntimeConfigManager.
- ProcessingResult DTO consolidated into MetricsManager.ProcessingResult.
- File utility methods (discoverPartitions, getFileBatches, etc.) moved
  from YAFPF.java to org.gamma.util.FileUtils.
- Concurrency utility methods (createPlatformThreadFactory,
  shutdownExecutorService, etc.) moved from YAFPF.java to
  org.gamma.util.ConcurrencyUtils.

Incomplete Refactoring & Known Issues:
- The `Status` enum and `HasStatus` interface were nested within
  `MetricsManager`. This required updating all dependent files.
- While many files were updated (YAFPF, processors, some handlers,
  StatusHelper, FileInfo), several others still need to be updated to use
  `MetricsManager.Status` and `MetricsManager.HasStatus`.
- **Your codebase currently does not compile due to these pending updates.**
- Known files requiring updates for `Status`/`HasStatus`:
    - `org.gamma.metrics.LoadMetrics` (needs to implement
      `MetricsManager.HasStatus` and use `MetricsManager.Status`)
    - `org.gamma.metrics.PollInfo` (use `MetricsManager.Status`)
    - `org.gamma.processing.SimulatedFileLoader` (use `MetricsManager.Status`)
    - `org.gamma.processing.AbstractBatchFileParser` (check for `Status` usage)
    - `org.gamma.datasources.AIRFileParser` (check for `Status` usage)
- Further iterative compilation and fixing would be needed to resolve these.

The OOD enhancement of removing the YAFPF instance dependency from processor
classes is also not yet complete due to these cascading type issues.
