+++ src/main/java/org/gamma/config/AggregationDetailItem.java
@@ -1,0 +1,9 @@
+package org.gamma.config;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record AggregationDetailItem(
+    @JsonProperty("input_field") String inputField,
+    String function
+) {
+}
