+++ src/main/java/org/gamma/config/ConfigManager.java
@@ -1,6 +1,6 @@
 package org.gamma.config;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
+// import com.fasterxml.jackson.annotation.JsonProperty; // No longer needed directly
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
@@ -8,129 +8,46 @@ import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.Executors;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicInteger;
+// import java.util.List; // AppConfig might use it, but ConfigManager itself doesn't directly. Will be caught by compiler if needed by AppConfig's direct usage here.
+// import java.util.Map;
+// import java.util.concurrent.ConcurrentHashMap;
+// import java.util.concurrent.atomic.AtomicBoolean;
+// import java.util.concurrent.atomic.AtomicInteger;
+// Executors, ScheduledExecutorService, TimeUnit were only for the main method, so they remain removed.
 import java.util.logging.ConsoleHandler;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.logging.SimpleFormatter;
 
-record DatabaseConnectionItem(String connectionName, String type, String dbUrl, String dbUser, String dbPass,
-                              Integer maxPool, Long idleTimeout, Long connTimeout, Long maxLifetime) {
-}
+// DatabaseConnectionItem moved to its own file
 
-record PipelineDetailsConfig(String pipelineName, List<SourceDetailsItem> sourcesDetails, List<OutputItem> outputs) {
-}
+// PipelineDetailsConfig moved to its own file
 
-record SourceDetailsItem(String sourceId, String sourceMetadata, String sourceFields,
-                         ValidationRulesItem validationRules) {
-}
+// SourceDetailsItem moved to its own file
 
-record FieldValidationConfigItem(String fieldName, List<ValidationExpressionItem> validations) {
-}
+// FieldValidationConfigItem moved to its own file
 
-record ValidationExpressionItem(String expression, String errorMessage, String actionOnFailure) {
-}
+// ValidationExpressionItem moved to its own file
 
-record TransformationItem(String transformationName, String type, String expression) {
-}
+// TransformationItem moved to its own file
 
-record AggregationDetailItem(@JsonProperty("input_field") String inputField, String function) {
-}
+// AggregationDetailItem moved to its own file
 
-record AdjustableParametersConfig(AdjustableParameterDetailConfig numThreads,
-                                  AdjustableParameterDetailConfig batchSize) {
-}
+// AdjustableParametersConfig moved to its own file
 
-record AdjustableParameterDetailConfig(Integer min, Integer max, Integer incrementStep, Integer decrementStep,
-                                       Double incrementFactor, Double decrementFactor) {
-}
+// AdjustableParameterDetailConfig moved to its own file
 
-record TuningRuleConfig(String ruleName, String metric, String condition, Double threshold, String conditionLogic,
-                        List<SingleConditionConfig> conditions, Integer consecutiveBreachesToTrigger,
-                        RuleActionConfig action) {
-}
+// TuningRuleConfig moved to its own file
 
-record SingleConditionConfig(String metric, String condition, Double threshold) {
-}
+// SingleConditionConfig moved to its own file
 
-record RuleActionConfig(String type, Integer durationMinutes, RuleActionParamsConfig params,
-                        RuleActionConfig interimAction, String reason) {
-}
+// RuleActionConfig moved to its own file
 
-record RuleActionParamsConfig(Integer numThreadsIncrement, Integer numThreadsDecrement, String targetNumThreads,
-                              Double batchSizeIncrementFactor, Double batchSizeDecrementFactor,
-                              Double targetBatchSizeFactor) {
-}
+// RuleActionParamsConfig moved to its own file
 
-record SourceRuntimeState(String sourceId, AtomicInteger currentNumThreads, AtomicInteger currentBatchSize,
-                          AtomicBoolean currentActive) {
-    public SourceRuntimeState(SourceItem sourceConfig) { // Takes simplified SourceItem
-        this(sourceConfig.sourceId(), new AtomicInteger(sourceConfig.numThreads() != null ? sourceConfig.numThreads() : 1),
-                new AtomicInteger(sourceConfig.batchSize() != null ? sourceConfig.batchSize() : 10),
-                new AtomicBoolean(sourceConfig.active()));
-    }
-}
+// SourceRuntimeState and PipelineRuntimeState moved to RuntimeConfigManager.java
 
-record PipelineRuntimeState(String pipelineName, AtomicBoolean currentActive,
-                            Map<String, SourceRuntimeState> sourceStates, AutoTuningConfig tuningConfig,
-                            PipelineDetailsConfig detailsConfig,
-                            Map<String, AtomicInteger> ruleConsecutiveBreaches, long pipelineCooldownUntilTimestamp) {
-    public PipelineRuntimeState(EtlPipelineItem pipelineConfig, PipelineDetailsConfig detailsConfig) {
-        this(pipelineConfig.pipelineName(), new AtomicBoolean(pipelineConfig.active()),
-                new ConcurrentHashMap<>(), pipelineConfig.autoTuning(), detailsConfig, // Store the loaded details
-                new ConcurrentHashMap<>(), 0L);
-        if (pipelineConfig.sources() != null) {
-            for (SourceItem sourceCoreConfig : pipelineConfig.sources())
-                this.sourceStates.put(sourceCoreConfig.sourceId(), new SourceRuntimeState(sourceCoreConfig));
-        }
-        if (this.tuningConfig != null && this.tuningConfig.tuningRules() != null) {
-            for (TuningRuleConfig rule : this.tuningConfig.tuningRules())
-                this.ruleConsecutiveBreaches.put(rule.ruleName(), new AtomicInteger(0));
-        }
-    }
-}
-
-class MetricSimulator {
-    private static final Logger LOGGER = Logger.getLogger(MetricSimulator.class.getName());
-    private double currentCpuLoad = 30.0;
-    private double currentIoWait = 5.0;
-    private int currentQueueLength = 50;
-    private double currentErrorRate = 1.0;
-
-    public Map<String, Double> getMetricsForPipeline(String pipelineName) {
-        currentCpuLoad += (Math.random() - 0.5) * 20;
-        currentCpuLoad = Math.max(5.0, Math.min(95.0, currentCpuLoad));
-        currentIoWait += (Math.random() - 0.5) * 5;
-        currentIoWait = Math.max(1.0, Math.min(40.0, currentIoWait));
-        if (Math.random() < 0.3) {
-            currentQueueLength += (int) ((Math.random() - 0.4) * 100);
-            currentQueueLength = Math.max(0, Math.min(500, currentQueueLength));
-        }
-        if (Math.random() < 0.1)
-            currentErrorRate = Math.random() * 15;
-        else if (Math.random() < 0.3)
-            currentErrorRate = Math.random() * 2;
-
-        currentErrorRate = Math.max(0.0, Math.min(20.0, currentErrorRate));
-
-        Map<String, Double> metrics = new ConcurrentHashMap<>();
-        metrics.put("systemCpuLoadPercent", currentCpuLoad);
-        metrics.put("systemIoWaitPercent", currentIoWait);
-        metrics.put("pipelineInputQueueLength", (double) currentQueueLength);
-        metrics.put("pipelineErrorRatePercentLastHour", currentErrorRate);
-        LOGGER.fine(String.format("Simulated metrics for pipeline '%s': CPU=%.1f%%, IO_Wait=%.1f%%, Queue=%d, ErrorRate=%.1f%%",
-                pipelineName, currentCpuLoad, currentIoWait, currentQueueLength, currentErrorRate));
-        return metrics;
-    }
-}
-
+// MetricSimulator class has been moved to src/main/java/org/gamma/config/MetricSimulator.java (corrected path)
 
 public class ConfigManager {
     private static final Logger APP_LOGGER = Logger.getLogger(ConfigManager.class.getName());
@@ -143,79 +60,14 @@ public class ConfigManager {
         Logger rootLogger = Logger.getLogger("");
         rootLogger.addHandler(handler);
         rootLogger.setLevel(Level.INFO);
-        Logger.getLogger(RuntimeConfigManager.class.getName()).setLevel(Level.INFO);
-        Logger.getLogger(AutoTuner.class.getName()).setLevel(Level.INFO);
-        Logger.getLogger(MetricSimulator.class.getName()).setLevel(Level.FINE);
+        // Logger.getLogger(RuntimeConfigManager.class.getName()).setLevel(Level.INFO); // RuntimeConfigManager related logic moved
+        // Logger.getLogger(AutoTuner.class.getName()).setLevel(Level.INFO); // AutoTuner related logic moved
+        // Logger.getLogger(MetricSimulator.class.getName()).setLevel(Level.FINE); // MetricSimulator moved
     }
 
     static AppConfig appConfig;
 
-    public static void main(String[] args) {
-        ObjectMapper frameworkObjectMapper = new ObjectMapper(new YAMLFactory());
-        frameworkObjectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-
-        String configPath = "conf/config.yaml";
-//        if (baseConfigPath == null) baseConfigPath = "."; // Default to current dir if parent is null
-
-        try {
-            APP_LOGGER.info("Reading main framework configuration from: " + configPath);
-            appConfig = frameworkObjectMapper.readValue(new File(configPath), AppConfig.class);
-            APP_LOGGER.info("Successfully read initial main framework configuration.");
-
-            Path baseConfigPath = Path.of("conf"); // Get directory of main config
-            RuntimeConfigManager rtConfigManager = new RuntimeConfigManager(appConfig, baseConfigPath);
-            AutoTuner autoTuner = new AutoTuner(rtConfigManager);
-
-            autoTuner.start();
-
-            try (ScheduledExecutorService service = Executors.newSingleThreadScheduledExecutor()) {
-                service.scheduleAtFixedRate(() -> {
-                    if (appConfig.etlPipelines() == null) return;
-                    for (EtlPipelineItem pipelineItem : appConfig.etlPipelines()) {
-                        String pipelineName = pipelineItem.pipelineName();
-                        // Check if pipeline was successfully loaded by RuntimeConfigManager
-                        if (rtConfigManager.getPipelineRuntimeState(pipelineName) == null) {
-                            APP_LOGGER.warning("ETL Module for Pipeline '" + pipelineName + "': Skipped as its configuration was not fully loaded.");
-                            continue;
-                        }
-
-                        if (rtConfigManager.isPipelineActive(pipelineName)) {
-                            if (pipelineItem.sources() != null && !pipelineItem.sources().isEmpty()) {
-                                String sourceId = pipelineItem.sources().getFirst().sourceId();
-                                int numThreads = rtConfigManager.getCurrentNumThreads(pipelineName, sourceId);
-                                int batchSize = rtConfigManager.getCurrentBatchSize(pipelineName, sourceId);
-                                APP_LOGGER.info(String.format("ETL Module for Pipeline '%s', Source '%s': Currently using numThreads=%d, batchSize=%d",
-                                        pipelineName, sourceId, numThreads, batchSize));
-
-                                // Example: Accessing detailed config
-                                rtConfigManager.getValidationRules(pipelineName, sourceId).ifPresent(rules ->
-                                        APP_LOGGER.fine("  Validation rules defaultMaxFieldSize: " + rules.defaultMaxFieldSize()));
-                                rtConfigManager.getOutputs(pipelineName).ifPresent(outputs ->
-                                        APP_LOGGER.fine("  Number of outputs defined: " + outputs.size()));
-                            }
-                        } else
-                            APP_LOGGER.info(String.format("ETL Module for Pipeline '%s': Pipeline is currently INACTIVE.", pipelineName));
-
-                    }
-                    System.out.println("----------------------------------------------------");
-                }, 10, 20, TimeUnit.SECONDS); // Check every 20 seconds
-
-
-                Thread.sleep(TimeUnit.MINUTES.toMillis(5));
-
-                APP_LOGGER.info("Shutting down ETL simulation and performance tuner...");
-                autoTuner.stop();
-                service.shutdownNow();
-            }
-            APP_LOGGER.info("Application finished.");
-
-        } catch (IOException e) {
-            APP_LOGGER.log(Level.SEVERE, "Error processing YAML configuration: " + e.getMessage(), e);
-        } catch (InterruptedException e) {
-            APP_LOGGER.log(Level.INFO, "Application interrupted.");
-            Thread.currentThread().interrupt();
-        }
-    }
+    // main method has been moved to src/test/java/org/gamma/config/simulation/ConfigSimulation.java
 
     public static AppConfig getConfig() throws IOException {
         if (appConfig == null) {
