+++ src/main/java/org/gamma/config/FieldValidationConfigItem.java
@@ -1,0 +1,12 @@
+package org.gamma.config;
+
+import java.util.List;
+
+// ValidationExpressionItem is expected to be in the same package
+// or imported if moved to its own file later.
+
+public record FieldValidationConfigItem(
+    String fieldName,
+    List<ValidationExpressionItem> validations
+) {
+}
