+++ src/main/java/org/gamma/config/SingleConditionConfig.java
@@ -1,0 +1,10 @@
+package org.gamma.config;
+
+// No special imports needed for this record beyond standard Java types.
+
+public record SingleConditionConfig(
+    String metric,
+    String condition,
+    Double threshold
+) {
+}
