+++ src/main/java/org/gamma/config/ValidationExpressionItem.java
@@ -1,0 +1,10 @@
+package org.gamma.config;
+
+// No special imports needed for this record beyond standard Java types.
+
+public record ValidationExpressionItem(
+    String expression,
+    String errorMessage,
+    String actionOnFailure
+) {
+}
